backend/requirements/base.txt
Django==3.0
django-environ==0.4.5
psycopg2-binary==2.8.4
argon2-cffi==19.2.0
flake8==3.7.9
djangorestframework==3.11.0
djangorestframework_simplejwt==4.4.0
django-cors-headers==3.2.0
sphinx==2.3.1
sphinx-rtd-theme==0.4.3
